feat(notifications): create HOC to handle notifications
